app/biowl/requirements.txt
Flask
Flask-HTTPAuth
Flask-RESTful
Jinja2
MarkupSafe
Werkzeug
itsdangerous
six
pyparsing
future
hdfs==2.0.15
biopython
bioblend